Basic start to the radial plot.  Computes distance and bearing for all data points relative to an origin (currently Davis) and plots on a logarithmic polar plot.
Still needs to scale points based on size and color based on class as well as any other filtering, but since it processes the data in a loop it should be a small matter to check the filter element and then skip the data being filtered out during the calculations.  linelogcomp.png shows the difference between a linear plot and a logarithmic plot.